gpg: minor renames and code refactoring
